Expect ValueError to be raised
According to the pytest docs, `xfail` should be used for tests,
which are expected to fail due to a feature not being implemented,
or a bug not yet fixed.

When the expectation of a test is an error to be raised,
`pytest.raises` should be used to check this expectation.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>