Raise exception for validate
Problem:
Currently, an exception is not raised if the `validate` method is
called, even when there are errors and the validation fails.
Consequently, the cli module does not catch an exception and the courier
does not modify its exit code from 0 when the `validate` method is used.
`operator-courier` should return a non zero exit code when the input
is invalid.

Fix:
Raise a ValueError in the build_and_verify method when the bundle isn't
valid.
Remove the empty bundle check in the push method, since the underlying
build_and_verify method will raise the necessary exception

Fixes https://github.com/operator-framework/operator-courier/issues/46V