Use argparse subparsers for subcommands
Makes argument parsing more resilient, by handling
subcommands with argparse provided and recommended [subparsers][1],
instead of the current custom manipulation of argv.

Removes the need to manually update and format the top-level
help message: this is going to be composed from the help strings
of the subcommands by argparse.

Fixes the issue of subcommand's usage string not displaying the
subcommand.

Removes the need to explicitly check for incorrect subcommands.

Enables extending the CLI with top-level optional arguments, without the
need of adjusting argv parsing.

[1]: https://docs.python.org/3/library/argparse.html#sub-commands

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>