Add Optional Output File
Problem:
Currently, when running the verify or push commands there is no easy way
to actually inspect the bundle that is created -- it is a temp file that
is discarded when the command execution completes. Ideally, it would
be useful to optionally dump the bundle yaml to a specified file.

Solution:
Add optional output parameter that dumps the bundle to specified file.

This addresses issue
https://github.com/operator-framework/operator-courier/issues/76