Remove SemVer requirement on nested folders
Problem:
Technically, folder names need not match the version name in a manifest
directory. Enforcing this is just convention. We should remove this.

Solution:
Remove the function and lambda from the verified_manifest class that
generates the dictionary needed to parse these files.

Additionally, update the tests to include a case when the folders are
not named with semver.

Also fix some directories that cannot be parsed. An artifact of this pr
is that flat directories now cannot include folders that should be
ignored. This will be addressed in a separate pr.