Add filename to warnings and errors
Problem:
Command line output logs are hard to read because, in the case where
there are many operator yamls, there is no obvious output as to which
file caused the error in question.

Solution:
Add file references and relative file paths to the output logs.
Update build_bundle() to take the filename metadata so that the
validator can use the new metadata field